Rename SyncInstrument to SynchronousInstrument
