+++ api/src/main/java/io/opentelemetry/metrics/DoubleMeasure.java
@@ -70,7 +70,7 @@ public interface DoubleMeasure extends Measure<BoundDoubleMeasure> {
    * @since 0.1.0
    */
   @ThreadSafe
-  interface BoundDoubleMeasure extends SyncInstrument.BoundInstrument {
+  interface BoundDoubleMeasure extends SynchronousInstrument.BoundInstrument {
     /**
      * Records the given measurement, associated with the current {@code Context}.
      *
