+++ api/src/main/java/io/opentelemetry/metrics/LongMeasure.java
@@ -70,7 +70,7 @@ public interface LongMeasure extends Measure<BoundLongMeasure> {
    * @since 0.1.0
    */
   @ThreadSafe
-  interface BoundLongMeasure extends SyncInstrument.BoundInstrument {
+  interface BoundLongMeasure extends SynchronousInstrument.BoundInstrument {
     /**
      * Records the given measurement, associated with the current {@code Context}.
      *
