+++ api/src/main/java/io/opentelemetry/metrics/Measure.java
@@ -16,7 +16,7 @@
 
 package io.opentelemetry.metrics;
 
-import io.opentelemetry.metrics.SyncInstrument.BoundInstrument;
+import io.opentelemetry.metrics.SynchronousInstrument.BoundInstrument;
 import java.util.Map;
 import javax.annotation.concurrent.ThreadSafe;
 
@@ -27,7 +27,7 @@ import javax.annotation.concurrent.ThreadSafe;
  * @since 0.1.0
  */
 @ThreadSafe
-public interface Measure<H extends BoundInstrument> extends SyncInstrument<H> {
+public interface Measure<H extends BoundInstrument> extends SynchronousInstrument<H> {
 
   /** Builder class for {@link Measure}. */
   interface Builder extends Instrument.Builder {
