+++ sdk/src/main/java/io/opentelemetry/sdk/metrics/AbstractBoundInstrument.java
@@ -16,7 +16,7 @@
 
 package io.opentelemetry.sdk.metrics;
 
-import io.opentelemetry.metrics.SyncInstrument.BoundInstrument;
+import io.opentelemetry.metrics.SynchronousInstrument.BoundInstrument;
 import io.opentelemetry.sdk.metrics.aggregator.Aggregator;
 import java.util.concurrent.atomic.AtomicLong;
